API/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
librosa==0.10.1
numpy==1.24.3
python-multipart==0.0.6
tensorflow==2.13.0
matplotlib==3.7.1
Pillow==10.0.0